requirements.txt
PyYAML
RPi.GPIO==0.6.3
